use stored specs in compute-URLs
This commit adds the ability to store the specs
for a computation ---i.e. method-name, inputs,
outputs, and parameters--- in `.datalad`, and
reference it in the compute-URL. That keeps the
compute URLs short.

Note the the dataset version that is generated
by storing the specs is used as reference for
a computation.

If anything goes wrong in the computation, the
specs are still stored in the dataset.