fix: deactivate dirty check in provision-command
This commit disables the dirty check in the provisioning
command because `Dataset.status` does not work as expected
when executed in the context of a git annex special remote.

In this context it might detect a changed subdataset, which
should not happen. The reason for this is not clear yet.
Once this is resolved, this commit could be reverted.