fix: fix chained provision in Windows
This commit changes the way in which worktrees
are generated in Windows. In Windows, every provisioned
worktree will have a git config setting that identifies
it as a worktree and holds the source of the worktree.

Further provision calls, i.e. chained provision calls,
in this worktree will use the original source as
provision source. In addition they will explicitly
enable the datalad remake special remote in the provisioned
datasets.

This is necessary because `git worktree` does not work
together with `git annex` on a crippled file systems, which
means it does not work on most Windows systems.