doc: Start creating a basic docs structure
This removes generation of documentation for "commands" (no meaningful
docstrings) and instead adds an autosummarry for datalad.api (under
which these commands are available). This also adds an autosummary for
annexremotes. Both are moved into separate files, linked from the
index, because this way they end up as separate pages and separate
sidebar menus.

Doing it this way (datalad.api) requires that datalad is installed for
the documentation building. For developer experience (hatch env run
--env docs build) this requires installation dependencies to be set
properly.

Overall, the documentation structure tries to follow datalad-next as
much as possible.

One glaring omission is the CLI, arguably more important than the
Python API. However, datalad-next faces the same problem. We used to
rely on _datalad_buildsupport (contained in the extension codebase) to
generate manpages (also rst) for the commands. This is now gone, but
we are in a transition period with no direct replacement.