doc: Add documentation for file inputs used by remake
This adds a documentation of format and syntax of the files read or
written by remake: compute template, lists of inputs / outputs /
parameters, generated compute specification.

I did my best in collating / rephrasing the help message for datalad
make together with docstrings, and looking up the implementation if
necessary, but as-committed this could benefit from a second pair of
eyes.