ci: use optional dependencies instead of requirements-devel
The requirements-devel file was removed in favour of optional
dependencies in pyproject.toml. This changes a long-standing pattern
used across datalad packages in the docbuild CI action, but it should
not introduce any maintenance issues. In fact, it should make the
docbuild action ever so slightly faster, since it won't pull all the
development dependencies for testing.

We could take it one step further and use hatch (like most other
GitHub actions in this repo) but that does not seem necessary either.