ci: switch appveyor test setup to rely more on hatch & uv
This introduces the use of uv for package management (also git-annex,
thus replacing datalad-installer) and makes use of hatch to manage the
python versions used for testing (instead of appveyor's matrix). This
also means using 64 bit python on Windows, instead of 32 bit which
seems to have been the default.

Changes made here reflect similar changes done in datalad-core and
datalad-next, but they are a translation of the existing appveyor file
rather than a copy-paste from other projects.

The highest tested Python version is changed from 3.12 to 3.13. Tests
on MacOS and GNU/Linux are done on 3.9 & 3.13 while Windows is done on
3.13 only. Previously Windows was also done with one version and this
is related to build time limitation: things are slower on Windows, so
there is time to only test one if we try to have one appveyor job per
OS.

The changes should address (more or less directly) build issues
related to git-annex dependency (on Windows) and brew install (on
MacOS) while also improving cohesion with DataLad extensions mentioned
above.